pyproject.toml
[project]
name = "neohome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "textual>=4.0.0",
]
